.ipynb_checkpoints/requirements-checkpoint.txt
scikit_image
#git+https://gitlab.com/m1lhaus/blender-mathutils.git
tqdm==4.19.9
pandas
h5py
matplotlib
scipy
pyquaternion
opencv-python
cupy_cuda90==6.0.0
pykitti
numpy
tensorboardX
open3d-python==0.7.0.0
Pillow
scikit-image
torch==1.0.1.post2
torchvision==0.2.2
sacred==0.7.4
-e ./models/correlation_package
